HSEARCH-4701 Move *BackendMappingHints classes to packages *.reporting.spi/*.reporting.impl
Because that's where they should be in order to work properly with our
module-info definitions.
