+++ engine/src/main/java/org/hibernate/search/engine/backend/reporting/spi/BackendMappingHints.java
@@ -4,9 +4,8 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.orm.logging.impl;
+package org.hibernate.search.engine.backend.reporting.spi;
 
-import org.hibernate.search.engine.backend.mapping.spi.BackendMappingHints;
 import org.hibernate.search.util.common.logging.impl.MessageConstants;
 
 import org.jboss.logging.Messages;
@@ -14,12 +13,11 @@ import org.jboss.logging.annotations.Message;
 import org.jboss.logging.annotations.MessageBundle;
 
 @MessageBundle(projectCode = MessageConstants.PROJECT_CODE)
-public interface HibernateOrmMappingHints extends BackendMappingHints {
+public interface BackendMappingHints {
 
-	HibernateOrmMappingHints INSTANCE = Messages.getBundle( HibernateOrmMappingHints.class );
+	BackendMappingHints NONE = Messages.getBundle( BackendMappingHints.class );
 
-	@Override
-	@Message(value = "This is likely a bug, as Hibernate ORM entities should always be loadable from the database.")
+	@Message(value = "")
 	String noEntityProjectionAvailable();
 
 }
