+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/reporting/impl/HibernateOrmMappingHints.java
@@ -4,8 +4,9 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.engine.backend.mapping.spi;
+package org.hibernate.search.mapper.orm.reporting.impl;
 
+import org.hibernate.search.engine.backend.reporting.spi.BackendMappingHints;
 import org.hibernate.search.util.common.logging.impl.MessageConstants;
 
 import org.jboss.logging.Messages;
@@ -13,11 +14,12 @@ import org.jboss.logging.annotations.Message;
 import org.jboss.logging.annotations.MessageBundle;
 
 @MessageBundle(projectCode = MessageConstants.PROJECT_CODE)
-public interface BackendMappingHints {
+public interface HibernateOrmMappingHints extends BackendMappingHints {
 
-	BackendMappingHints NONE = Messages.getBundle( BackendMappingHints.class );
+	HibernateOrmMappingHints INSTANCE = Messages.getBundle( HibernateOrmMappingHints.class );
 
-	@Message(value = "")
+	@Override
+	@Message(value = "This is likely a bug, as Hibernate ORM entities should always be loadable from the database.")
 	String noEntityProjectionAvailable();
 
 }
